% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusnames.R
\name{censusnames}
\alias{censusnames}
\title{Frequencies of Croatian personal names/surnames}
\usage{
censusnames(firstname, surname)
}
\arguments{
\item{firstname, surname}{Vectors of names.}
}
\value{
Data frame with frequencies of names. Frequencies less than 10 are not available and will be returned as missing values.
}
\description{
Function \code{censusnames} fetches the frequencies of names from \url{https://www.dzs.hr/hrv/censuses/census2011/results/censusnames.htm}.
}
\details{
Only first name or surname can be supplied, or both to fetch full name frequency. When supplying multiple first names and surnames, both vectors should be of the same length or one should be a unit vector and it will be replicated (with a warning). Queries are not case sensitive. Leading/trailing whitespace is trimmed and only unique names are retained.
}
\note{
When using this data, please acknowledge the source! For more info visit CBS page: \url{https://www.dzs.hr/default_e.htm}.
}
\examples{
\dontrun{
# First name
censusnames("Ana")

# Full name
censusnames("Ana","Fabijanec")

# Multiple first names, same surname
censusnames(c("Ana","Ivana"), "Fabijanec")

}

}
