% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrcensusnames-package.R, R/hrcensusnames.R
\docType{package}
\name{hrcensusnames}
\alias{hrcensusnames}
\alias{hrcensusnames-package}
\title{Frequencies of Croatian personal names/surnames}
\usage{
hrcensusnames(firstname, surname)
}
\arguments{
\item{firstname, surname}{Vectors of names.}
}
\value{
Data frame with frequencies of names. Frequencies less than 10 are not available and will be returned as missing values.
}
\description{
Croatian bureau of statistics (CBS) gives access to frequencies of personal names/surnames based on data from official census. This package offers functionality to fetch the frequencies and make it available for analysis in R.

Function \code{hrcensusnames} fetches the frequencies of names from \url{https://www.dzs.hr/hrv/censuses/census2011/results/censusnames.htm}.
}
\details{
Only first name or surname can be supplied, or both to fetch full name frequency. When supplying multiple first names and surnames, both vectors should be of the same length or one should be a unit vector and it will be replicated (with a warning). Queries are not case sensitive. Leading/trailing whitespace is trimmed and only unique names are retained.
}
\note{
When using this data, please acknowledge the source! For more info visit CBS page: \url{https://www.dzs.hr/default_e.htm}.
}
\examples{
\dontrun{
# First name
hrcensusnames("Ana")

# Full name
hrcensusnames("Ana","Fabijanec")

# Multiple first names, same surname
hrcensusnames(c("Ana","Ivana"), "Fabijanec")

}

}
